Feature/module implementation v2

+++ database/network_plugin.go
@@ -1,9 +1,13 @@
 package database
 
 import (
+	"encoding/json"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 	"github.com/NubeIO/rubix-os/api"
+	"github.com/NubeIO/rubix-os/constants"
+	"github.com/NubeIO/rubix-os/module/common"
 	"github.com/NubeIO/rubix-os/src/client"
+	"strings"
 	"time"
 )
 
@@ -16,13 +20,29 @@ func (d *GormDatabase) CreateNetworkPlugin(body *model.Network) (network *model.
 		}
 		return
 	}
+
 	body.CommonFault.InFault = true
 	body.CommonFault.MessageLevel = model.MessageLevel.NoneCritical
 	body.CommonFault.MessageCode = model.CommonFaultCode.PluginNotEnabled
 	body.CommonFault.Message = model.CommonFaultMessage.PluginNotEnabled
 	body.CommonFault.LastFail = time.Now().UTC()
 	body.CommonFault.LastOk = time.Now().UTC()
-	// if plugin like bacnet then call the api direct on the plugin as the plugin knows best how to add a point to keep things in sync
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return nil, moduleNotFoundError(pluginName)
+		}
+		bytes, _ := json.Marshal(body)
+		bytes, err = module.Post(common.NetworksURL, bytes)
+		if err != nil {
+			return nil, err
+		}
+		var net *model.Network
+		_ = json.Unmarshal(bytes, &net)
+		return net, nil
+	}
+
 	cli := client.NewLocalClient()
 	network, err = cli.CreateNetworkPlugin(body, pluginName)
 	if err != nil {
@@ -43,6 +63,22 @@ func (d *GormDatabase) UpdateNetworkPlugin(uuid string, body *model.Network) (ne
 		}
 		return
 	}
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return nil, moduleNotFoundError(pluginName)
+		}
+		bytes, _ := json.Marshal(body)
+		bytes, err = module.Patch(common.NetworksURL, uuid, bytes)
+		if err != nil {
+			return nil, err
+		}
+		var net *model.Network
+		_ = json.Unmarshal(bytes, &net)
+		return net, nil
+	}
+
 	cli := client.NewLocalClient()
 	if err != nil {
 		return nil, err
@@ -67,6 +103,19 @@ func (d *GormDatabase) DeleteNetworkPlugin(uuid string) (ok bool, err error) {
 		}
 		return
 	}
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return false, moduleNotFoundError(pluginName)
+		}
+		_, err = module.Delete(common.NetworksURL, uuid)
+		if err != nil {
+			return false, err
+		}
+		return true, nil
+	}
+
 	cli := client.NewLocalClient()
 	ok, err = cli.DeleteNetworkPlugin(network, pluginName)
 	if err != nil {
