+++ database/point_plugin.go
@@ -1,10 +1,14 @@
 package database
 
 import (
+	"encoding/json"
 	"github.com/NubeIO/nubeio-rubix-lib-models-go/pkg/v1/model"
 	"github.com/NubeIO/rubix-os/api"
+	"github.com/NubeIO/rubix-os/constants"
+	"github.com/NubeIO/rubix-os/module/common"
 	"github.com/NubeIO/rubix-os/src/client"
 	"github.com/NubeIO/rubix-os/utils/boolean"
+	"strings"
 	"time"
 )
 
@@ -23,13 +27,29 @@ func (d *GormDatabase) CreatePointPlugin(body *model.Point) (point *model.Point,
 		point, err = d.UpdatePoint(point.UUID, point)
 		return
 	}
+
 	body.CommonFault.MessageLevel = model.MessageLevel.NoneCritical
 	body.CommonFault.MessageCode = model.CommonFaultCode.PluginNotEnabled
 	body.CommonFault.Message = model.CommonFaultMessage.PluginNotEnabled
 	body.CommonFault.LastFail = time.Now().UTC()
 	body.CommonFault.LastOk = time.Now().UTC()
 	body.CommonFault.InFault = true
-	// if plugin like bacnet then call the api direct on the plugin as the plugin knows best how to add a point to keep things in sync
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return nil, moduleNotFoundError(pluginName)
+		}
+		bytes, _ := json.Marshal(body)
+		bytes, err = module.Post(common.PointsURL, bytes)
+		if err != nil {
+			return nil, err
+		}
+		var pnt *model.Point
+		_ = json.Unmarshal(bytes, &pnt)
+		return pnt, nil
+	}
+
 	cli := client.NewLocalClient()
 	point, err = cli.CreatePointPlugin(body, pluginName)
 	if err != nil {
@@ -52,6 +72,22 @@ func (d *GormDatabase) UpdatePointPlugin(uuid string, body *model.Point) (point 
 		}
 		return
 	}
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return nil, moduleNotFoundError(pluginName)
+		}
+		bytes, _ := json.Marshal(body)
+		bytes, err = module.Patch(common.PointsURL, uuid, bytes)
+		if err != nil {
+			return nil, err
+		}
+		var pnt *model.Point
+		_ = json.Unmarshal(bytes, &pnt)
+		return pnt, nil
+	}
+
 	cli := client.NewLocalClient()
 	point, err = cli.UpdatePointPlugin(body, pluginName)
 	if err != nil {
@@ -72,6 +108,21 @@ func (d *GormDatabase) WritePointPlugin(uuid string, body *model.PointWriter) (p
 			return nil, err
 		}
 		return
+	}
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return nil, moduleNotFoundError(pluginName)
+		}
+		bytes, _ := json.Marshal(body)
+		bytes, err = module.Patch(common.PointsWriteURL, uuid, bytes)
+		if err != nil {
+			return nil, err
+		}
+		var pnt *model.Point
+		_ = json.Unmarshal(bytes, &pnt)
+		return pnt, nil
 	}
 
 	cli := client.NewLocalClient()
@@ -99,6 +150,19 @@ func (d *GormDatabase) DeletePointPlugin(uuid string) (ok bool, err error) {
 		}
 		return
 	}
+
+	if strings.HasPrefix(pluginName, constants.ModulePrefix) {
+		module := d.Modules[pluginName]
+		if module == nil {
+			return false, moduleNotFoundError(pluginName)
+		}
+		_, err = module.Delete(common.PointsURL, uuid)
+		if err != nil {
+			return false, err
+		}
+		return true, nil
+	}
+
 	cli := client.NewLocalClient()
 	ok, err = cli.DeletePointPlugin(point, pluginName)
 	if err != nil {
